requirements.txt
amqp==2.5.2
anyjson==0.3.3
Babel==2.8.0
billiard==3.6.1.0
celery==4.4.0
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.7
django-allauth==0.41.0
django-appconf==1.0.3
django-countries==5.5
django-crispy-forms==1.8.1
django-phone-field==1.8.0
django-phonenumber-field==4.0.0
idna==2.8
gunicorn
importlib-metadata==1.3.0
kombu==4.6.7
more-itertools==8.0.2
oauthlib==3.1.0
phonenumbers==8.11.1
Pillow==7.0.0
psycopg2==2.8.4
python3-openid==3.1.0
pytz==2019.3
redis==3.3.11
requests==2.22.0
requests-oauthlib==1.3.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
vine==1.3.0
zipp==0.6.0
